feat: Add the relationship between the tables restaurants and users.
As restaurants have a manager, it is necessary to save the user ID of this manager in the restaurants table.

A migration was created to add the relationship between the restaurants and users tables. The 'manager_id' field was added to the 'restaurants' table to save the ID of the user who will be the manager of the restaurant.